+++ libngpu/src/include/ngpu/ngpu.h
@@ -1104,6 +1104,7 @@ struct ngpu_pgcraft_texture {
     int writable;
     enum ngpu_format format;
     int clamp_video;
+    int premult;
     /*
      * Just like the other types (uniforms, blocks, attributes), this field
      * exists in order to be transmitted to the pipeline (through the